feat: populate collections and mark as unpopulated
So that the collection fields will be an array of primary keys in the
JSON response and each table only need to be queried once.

While the old approach is to init the collections of each entity one
by one, which make one table be queried for more than once and
causes performance problems.